+++ pkg/generator/generator.go
@@ -17,11 +17,13 @@
 package generator
 
 import (
+	"bytes"
 	"text/template"
 
-	"github.com/loopholelabs/frpc-go/internal/utils"
 	"github.com/loopholelabs/frpc-go/internal/version"
 	"github.com/loopholelabs/frpc-go/templates"
+	"github.com/loopholelabs/polyglot-go/pkg/generator"
+	"github.com/loopholelabs/polyglot-go/pkg/utils"
 	"google.golang.org/protobuf/compiler/protogen"
 	"google.golang.org/protobuf/proto"
 	"google.golang.org/protobuf/types/pluginpb"
@@ -41,13 +43,6 @@ func init() {
 		"Counter":            utils.Counter,
 		"FirstLowerCase":     utils.FirstLowerCase,
 		"FirstLowerCaseName": utils.FirstLowerCaseName,
-		"FindValue":          findValue,
-		"GetKind":            getKind,
-		"GetLUTEncoder":      getLUTEncoder,
-		"GetLUTDecoder":      getLUTDecoder,
-		"GetEncodingFields":  getEncodingFields,
-		"GetDecodingFields":  getDecodingFields,
-		"GetKindLUT":         getKindLUT,
 		"GetServerFields":    getServerFields,
 	}).ParseFS(templates.FS, "*"))
 }
@@ -76,11 +71,40 @@ func (g *Generator) Generate(req *pluginpb.CodeGeneratorRequest) (res *pluginpb.
 		return nil, err
 	}
 
+	var tplBuffer bytes.Buffer
+	if err := templ.ExecuteTemplate(&tplBuffer, "customEncode.templ", nil); err != nil {
+		return nil, err
+	}
+	customEncode := tplBuffer.String()
+	tplBuffer.Reset()
+
+	if err := templ.ExecuteTemplate(&tplBuffer, "customDecode.templ", nil); err != nil {
+		return nil, err
+	}
+	customDecode := tplBuffer.String()
+	tplBuffer.Reset()
+
+	if err := templ.ExecuteTemplate(&tplBuffer, "customFields.templ", nil); err != nil {
+		return nil, err
+	}
+	customFields := tplBuffer.String()
+
+	gen := generator.New()
+	gen.CustomEncode = func() string {
+		return customEncode
+	}
+	gen.CustomDecode = func() string {
+		return customDecode
+	}
+	gen.CustomFields = func() string {
+		return customFields
+	}
+
 	for _, f := range plugin.Files {
 		if !f.Generate {
 			continue
 		}
-		genFile := plugin.NewGeneratedFile(fileName(f.GeneratedFilenamePrefix), f.GoImportPath)
+		genFile := plugin.NewGeneratedFile(FileName(f.GeneratedFilenamePrefix), f.GoImportPath)
 
 		packageName := string(f.Desc.Package().Name())
 		if packageName == "" {
@@ -91,21 +115,35 @@ func (g *Generator) Generate(req *pluginpb.CodeGeneratorRequest) (res *pluginpb.
 
 		numMethods := 0
 		for i := 0; i < numServices; i++ {
-			numMethods += f.Desc.Services().Get(i).Methods().Len()
+			nM := f.Desc.Services().Get(i).Methods().Len()
+			numMethods += nM
 		}
 
-		err = templ.ExecuteTemplate(genFile, "base.templ", map[string]interface{}{
+		err = templ.ExecuteTemplate(genFile, "prebase.templ", map[string]interface{}{
 			"pluginVersion":   version.Version,
 			"sourcePath":      f.Desc.Path(),
 			"package":         packageName,
 			"requiredImports": requiredImports,
 			"serviceImports":  serviceImports,
 			"methodImports":   methodImports,
-			"enums":           f.Desc.Enums(),
-			"messages":        f.Desc.Messages(),
-			"services":        f.Desc.Services(),
 			"numServices":     numServices,
 			"numMethods":      numMethods,
+		})
+		if err != nil {
+			return nil, err
+		}
+
+		err = gen.ExecuteTemplate(genFile, f, packageName, false)
+		if err != nil {
+			return nil, err
+		}
+
+		err = templ.ExecuteTemplate(genFile, "base.templ", map[string]interface{}{
+			"enums":       f.Desc.Enums(),
+			"messages":    f.Desc.Messages(),
+			"services":    f.Desc.Services(),
+			"numServices": numServices,
+			"numMethods":  numMethods,
 		})
 		if err != nil {
 			return nil, err